Replace Save API with one based more cleanly on the flash chip API.
