+++ examples/test_sram.rs
@@ -4,7 +4,7 @@
 
 use gba::{fatal, info, time_this01, Color};
 use gba::io::display::{DisplayControlSetting, DisplayMode, DISPCNT};
-use gba::sram::*;
+use gba::save::*;
 use gba::vram::bitmap::Mode3;
 
 fn set_screen_color(r: u16, g: u16, b: u16) {
@@ -46,11 +46,14 @@ fn check_status<T>(r: Result<T, Error>) -> T {
 }
 
 fn do_write(mut pat: Pattern, shift: usize) -> Result<(), Error> {
-    let access = get_accessor();
-    let block_ct = access.len()? >> shift;
+    let access = SaveAccess::new()?;
+    let block_ct = access.len() >> shift;
     let mut buffer = [0; MAX_BLOCK_SIZE];
 
-    info!(" - Writing SRAM...");
+    info!(" - Clearing media...");
+    access.prepare_write(0..access.len())?;
+
+    info!(" - Writing media...");
     for i in 0..block_ct {
         for j in 0..(1 << shift) {
             buffer[j] = pat.next();
@@ -61,17 +64,17 @@ fn do_write(mut pat: Pattern, shift: usize) -> Result<(), Error> {
     Ok(())
 }
 fn do_verify(mut pat: Pattern, shift: usize) -> Result<(), Error> {
-    let access = get_accessor();
-    let block_ct = access.len()? >> shift;
+    let access = SaveAccess::new()?;
+    let block_ct = access.len() >> shift;
     let mut buffer = [0; MAX_BLOCK_SIZE];
 
-    info!(" - Validating SRAM...");
+    info!(" - Validating media...");
     for i in 0..block_ct {
         access.read(i << shift, &mut buffer[0..(1 << shift)])?;
         for j in 0..(1 << shift) {
             let cur = pat.next();
             assert!(buffer[j] == cur,
-                    "SRAM read does not match SRAM write: {} != {} @ 0x{:x}",
+                    "Read does not match earlier write: {} != {} @ 0x{:x}",
                     buffer[j], cur, i * 512 + j);
         }
     }
@@ -90,36 +93,32 @@ fn main(_argc: isize, _argv: *const *const u8) -> isize {
     // set a pattern to show that the ROM is working at all.
     set_screen_color(31, 31, 0);
 
-    // set the sram type
-    use_eeprom_512b();
+    // set the save type
+    use_flash_64k();
     set_timer_for_timeout(3);
 
-    // check some metainfo on the sram type
-    let access = get_accessor();
-    let len = check_status(access.len());
-    if len == 0 {
-        panic!("No SRAM accessor is installed.")
-    }
-    info!("SRAM type: {:?}", check_status(access.sram_type()));
-    info!("SRAM size: {} bytes", len);
+    // check some metainfo on the save type
+    let access = check_status(SaveAccess::new());
+    info!("Media info: {:?}", access.media_info());
+    info!("Media size: {} bytes", access.len());
     info!("");
 
-    // actually test the sram implementation
-    if len >= (1 << 12) {
-        info!("Testing incrementing pattern with 4 KiB blocks...");
+    // actually test the save implementation
+    if access.len() >= (1 << 12) {
+        info!("[ Incrementing, 4KiB blocks ]");
         print_time(time_this01!(check_status(do_write(Pattern(128, false), 12))));
         print_time(time_this01!(check_status(do_verify(Pattern(128, false), 12))));
 
-        info!("Testing random pattern with 4 KiB blocks...");
+        info!("[ Random, 4KiB blocks ]");
         print_time(time_this01!(check_status(do_write(Pattern(2000, true), 12))));
         print_time(time_this01!(check_status(do_verify(Pattern(2000, true), 12))));
     }
 
-    info!("Testing incrementing pattern with 512 byte blocks...");
+    info!("[ Incrementing, 0.5KiB blocks ]");
     print_time(time_this01!(check_status(do_write(Pattern(0, false), 9))));
     print_time(time_this01!(check_status(do_verify(Pattern(0, false), 9))));
 
-    info!("Testing random pattern with 512 byte blocks...");
+    info!("[ Random, 0.5KiB blocks ]");
     print_time(time_this01!(check_status(do_write(Pattern(1000, true), 9))));
     print_time(time_this01!(check_status(do_verify(Pattern(1000, true), 9))));
