--- src/sram/asm_read_byte.s
@@ -1,22 +1,0 @@
-@
-@ char SramReadByte(const char* offset);
-@
-    .thumb
-    .global SramReadByte
-    .thumb_func
-    .align 2
-SramReadByte:
-    ldr r1, =SramReadByteInner
-    bx r1
-
-    .pool
-    .section .data
-
-    .thumb
-    .thumb_func
-    .align 2
-SramReadByteInner:
-    ldrb r0, [r0]
-    mov pc, lr
-
-    .section .text
